refactor(user): 플레이어에게 수익 계산 역할 분배
