+++ src/main/java/blackjack/domain/game/PlayerProfits.java
@@ -1,7 +1,5 @@
 package blackjack.domain.game;
 
-import blackjack.domain.user.Dealer;
-import blackjack.domain.user.Player;
 import java.util.Collections;
 import java.util.List;
 
@@ -9,10 +7,8 @@ public class PlayerProfits {
 
     private final List<PlayerProfit> playerProfits;
 
-    public PlayerProfits(Dealer dealer, List<Player> players) {
-        playerProfits = players.stream()
-                .map(player -> makePlayerProfit(dealer, player))
-                .toList();
+    public PlayerProfits(List<PlayerProfit> playerProfits) {
+        this.playerProfits = playerProfits;
     }
 
     public int calculateDealerProfit() {
@@ -24,18 +20,5 @@ public class PlayerProfits {
 
     public List<PlayerProfit> getPlayerProfits() {
         return Collections.unmodifiableList(playerProfits);
-    }
-
-    private PlayerProfit makePlayerProfit(Dealer dealer, Player player) {
-        if (player.checkBlackjackWithInitialCard()) {
-            return PlayerProfit.createWhenPlayerBlackjackWithInitialCard(player);
-        }
-        if (dealer.isBust()) {
-            return PlayerProfit.createWhenDealerBust(player);
-        }
-        if (player.isBust()) {
-            return PlayerProfit.createWhenPlayerBust(player);
-        }
-        return PlayerProfit.createByWinningType(player, GameResultType.parse(player.getPoint(), dealer.getPoint()));
     }
 }
