+++ src/test/java/blackjack/domain/user/PlayerTest.java
@@ -1,17 +1,17 @@
 package blackjack.domain.user;
 
 import static blackjack.fixture.CardFixture.make;
-import static blackjack.fixture.PlayerFixture.createBlackJackWithFinalHand;
-import static blackjack.fixture.PlayerFixture.createBlackJackWithInitialHand;
-import static blackjack.fixture.PlayerFixture.createBust;
-import static blackjack.fixture.PlayerFixture.createLessThanBlackjack;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardShape;
 import blackjack.domain.card.CardValue;
+import blackjack.domain.game.GameResultType;
+import blackjack.domain.game.PlayerProfit;
 import blackjack.exception.ExceptionMessage;
+import blackjack.fixture.DealerFixture;
+import blackjack.fixture.PlayerFixture;
 import java.util.List;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
@@ -91,20 +91,67 @@ class PlayerTest {
         );
     }
 
-    @ParameterizedTest
-    @DisplayName("초기 카드만으로 블랙잭을 만족하는지 여부를 확인할 수 있다.")
-    @MethodSource()
-    void canCheckBlackjackWithInitialCard(Player player, boolean isBlackjackWithInitialCards) {
-        assertThat(player.checkBlackjackWithInitialCard())
-                .isEqualTo(isBlackjackWithInitialCards);
+    @Test
+    @DisplayName("플레이어가 초기카드로 블랙잭을 만들 경우 수익을 구할 수 있다.")
+    void canCreateWhenPlayerBlackjackWithInitialCard() {
+        Dealer dealer = DealerFixture.createBust();
+        Player player = PlayerFixture.createBlackJackWithInitialHand("플레이어");
+
+        PlayerProfit profit = player.calculateProfit(dealer);
+
+        int expectedProfit = GameResultType.WIN_WITH_INITIAL_HAND_BLACKJACK.calculateProfit(player.getBettingAmount());
+        assertThat(profit.getNickname()).isEqualTo("플레이어");
+        assertThat(profit.getProfit()).isEqualTo(expectedProfit);
     }
 
-    static Stream<Arguments> canCheckBlackjackWithInitialCard() {
+    @ParameterizedTest
+    @DisplayName("딜러가 버스트일 경우 플레이어의 수익을 구할 수 있다.")
+    @MethodSource()
+    void canCreateWhenDealerBust(Player player, GameResultType expectedType) {
+        Dealer dealer = DealerFixture.createBust();
+
+        PlayerProfit profit = player.calculateProfit(dealer);
+
+        int expectedProfit = expectedType.calculateProfit(player.getBettingAmount());
+        assertThat(profit.getNickname()).isEqualTo(player.getNickname());
+        assertThat(profit.getProfit()).isEqualTo(expectedProfit);
+    }
+
+    static Stream<Arguments> canCreateWhenDealerBust() {
         return Stream.of(
-                Arguments.of(createBlackJackWithInitialHand("플레이어"), true),
-                Arguments.of(createBlackJackWithFinalHand("플레이어"), false),
-                Arguments.of(createBust("플레이어"), false),
-                Arguments.of(createLessThanBlackjack("플레이어"), false)
+                Arguments.of(PlayerFixture.createBust("플레이어1"),
+                        GameResultType.WIN),
+                Arguments.of(PlayerFixture.createBlackJackWithInitialHand("플레이어2"),
+                        GameResultType.WIN_WITH_INITIAL_HAND_BLACKJACK),
+                Arguments.of(PlayerFixture.createBlackJackWithFinalHand("플레이어3"),
+                        GameResultType.WIN),
+                Arguments.of(PlayerFixture.createLessThanBlackjack("플레이어4"),
+                        GameResultType.WIN)
+        );
+    }
+
+    @ParameterizedTest
+    @DisplayName("승패에 따라 플레이어의 수익을 구할 수 있다.")
+    @MethodSource()
+    void canCreateByWinningType(Dealer dealer, Player player, GameResultType expectedType) {
+        PlayerProfit profit = player.calculateProfit(dealer);
+
+        int expectedProfit = expectedType.calculateProfit(player.getBettingAmount());
+        assertThat(profit.getNickname()).isEqualTo("플레이어");
+        assertThat(profit.getProfit()).isEqualTo(expectedProfit);
+    }
+
+    static Stream<Arguments> canCreateByWinningType() {
+        return Stream.of(
+                Arguments.of(
+                        DealerFixture.createBlackJackWithFinalHand(),
+                        PlayerFixture.createBlackJackWithFinalHand("플레이어"), GameResultType.DRAW),
+                Arguments.of(
+                        DealerFixture.createBlackJackWithFinalHand(),
+                        PlayerFixture.createLessThanBlackjack("플레이어"), GameResultType.LOSE),
+                Arguments.of(
+                        DealerFixture.createLessThanBlackjack(),
+                        PlayerFixture.createBlackJackWithFinalHand("플레이어"), GameResultType.WIN)
         );
     }
 }
